requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.5
cycler==0.10.0
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
drf-yasg==1.13.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mysqlclient==1.4.2.post1
nltk==3.4
numpy==1.16.0
pandas==0.24.0
Pillow==5.4.1
pycparser==2.19
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
ruamel.yaml==0.15.87
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
textblob==0.15.3
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
xlrd==1.2.0
